Hasharray, ...
hasharray as a replacement for hashmap, that is a generalization of a hashmap and a hashset. The entry structure is not defined in the header. The initialization of each entry is done by the caller of the function `Put`. That function only (al)locates an empty entry that could be filled. The caller could simply do nothing, so the hasharray will not contain the entry, because of the null empty representation. The null empty representation should not be created on an existing entry.
Currently only the template-header has been written, the template-source is coming soon