Removed alternative code
There was alternative code that was tested if it was more efficient. Through `#if - #else` and `#define` the code could be exchanged with another code, that has the same observable behavior, but might run faster/slower. The kept code is the best performing (on my machine) and the simplest code that I have came up with.