Memory management and related bugfixes
LAB_memory.h contains debugging versions of the allocation functions, which print all kinds of allocations and count each type. The final function LAB_DbgMemShow prints out those information. The checksum is all involved memory adresses xored together. When everything goes right, this value is 0. When one object does not get freed, the checksum is the address of this object. When more objects are not freed, the checksum is the memory addresses of those objects xored together.