The contents of each view chunk entry are now stored in an individual object
The contents of each view chunk entry are now stored in an individual object,
the view table is more likely to fit better in cache, but at a cost,
more memory allocations are performed and an additional level of indirection
is added. But the advantage is, that the addresses of these entries are now
stable, and the content of the entry is not moved around as much, when resizing
the hashtable buffer. Also it is now possible to add pointers to the entries
that point to neighboring entries (allowing faster checks, etc...)

Maybe the coordinates of the chunk entry could be moved back into the table
(reducing one level of indirection, when locating an element in the array)